+++ src/features/map/components/WeatherOverlay.tsx
@@ -1,9 +1,10 @@
-import { memo, useMemo } from "react";
+import { memo } from "react";
 import type { TodayWeatherResponse } from "../schemas/weather.schema";
 import styles from "./WeatherOverlay.module.css";
-import { toEmojiByCode, normalizeTemp } from "../utils/weather";
 import { useModalStore } from "@/common/hooks/useModalStore";
 import { WeatherModalContent } from "./WeatherModalContent";
+import { useWeatherButtonLabel } from "../hooks/useWeatherButtonLabel";
+import { WeatherButton } from "./WeatherButton";
 
 type Props = {
   weather: TodayWeatherResponse | null;
@@ -14,14 +15,7 @@ type Props = {
 function WeatherOverlayComponent({ weather, loading, error }: Props) {
   const open = useModalStore(state => state.open);
 
-  const buttonText = useMemo(() => {
-    if (loading) return "불러오는 중...";
-    if (error) return "오류";
-    if (!weather) return "날씨 없음";
-    const emoji = toEmojiByCode(weather.sky, weather.pty);
-    const temp = normalizeTemp(weather.tmp);
-    return `${emoji} ${temp}`;
-  }, [weather, loading, error]);
+  const buttonText = useWeatherButtonLabel(weather, loading, error);
 
   const handleOpen = () => {
     open(<WeatherModalContent weather={weather} loading={loading} error={error} />, {
@@ -31,15 +25,7 @@ function WeatherOverlayComponent({ weather, loading, error }: Props) {
 
   return (
     <div className={styles.container}>
-      <button
-        type="button"
-        className={styles.button}
-        disabled={loading || !!error}
-        onClick={handleOpen}
-        aria-expanded={false}
-      >
-        {buttonText}
-      </button>
+      <WeatherButton label={buttonText} disabled={loading || !!error} onClick={handleOpen} />
     </div>
   );
 }
