+++ src/features/map/components/index.ts
@@ -1,2 +1,3 @@
 export { WeatherOverlay } from "./WeatherOverlay";
 export { WeatherModalContent } from "./WeatherModalContent";
+export { WeatherButton } from "./WeatherButton";
